requirements.txt
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
blinker==1.7.0
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
channels==4.0.0
channels-redis==4.2.0
chardet==3.0.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
constantly==23.10.4
cryptography==42.0.5
daphne==4.1.0
dash==2.9.3
dash-bootstrap-components==1.5.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-mantine-components==0.12.1
dash_renderer==1.9.1
dash-table==5.0.0
defusedxml ==0.7.1
Django==5.0.3
django-allauth==0.61.1
django-plotly-dash==2.3.0
django-redis==5.4.0
dpd-components==0.1.0
filelock==3.13.1
Flask==3.0.2
fsspec==2024.3.1
google-auth==2.28.2
google-auth-oauthlib==1.2.0
googletrans==3.0.0
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2024.3.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
incremental==22.10.0
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
libretranslatepy==2.1.1
lxml==5.1.0
MarkupSafe==2.1.5
mpmath==1.3.0
msgpack==1.0.8
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
packaging==24.0
pandas==2.2.1
pillow==10.2.0
pip==24.0
plotly==5.20.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAudio==0.2.14
pycparser==2.21
PyJWT==2.8.0
pyOpenSSL==24.1.0
python-dateutil==2.9.0.post0
python3-openid==3.2.0
pytz==2024.1
redis==5.0.3
regex==2023.12.25
requests==2.31.0
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
service-identity==24.1.0
setuptools==69.2.0
six==1.16.0
sniffio==1.3.1
SpeechRecognition==3.10.1
sqlparse==0.4.4
sympy==1.12
tenacity==8.2.3
torch==2.2.1
tqdm==4.66.2
translate==3.6.1
Twisted==24.3.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.1
wheel==0.43.0
whitenoise==6.6.0
zope.interface==6.2
